Add workaround for Discord android/web bug
Currently there is a bug in Android and web where messages from the same
webhook are collapsed even if they have different info provided for the
"author" field.

To work around the issue we automatically create a second webhook (the
first one is made by the user) so we can alternate between webhooks
whenever the user changes. This is transparent on the Discord side,
except that the webhook appears in the list when the bridge is started.

The bridge automatically removes these webhooks on quit.

The issue on Discord's Trello: https://trello.com/c/6ZZTEBJw