transmitter: drop ineffassigns by returning bools
Previously we would use the webhook returned by getWebhook and
createWebhook just for comparisons against nil. Now these functions only
return a boolean value (and an error) indicating success.

getWebhook has been changed to freeWebhook to reflect this change.
freeWebhook might need tweaking as it returns a boolean value indicating
whether there is a free webhook to use, and doesn't return a bool
indicating if a webhook was actually freed up. This is significant when
freeWebhook returns true if a webhook for that channel already exists.