Add better puppet detection instead of detecting by suffix (#69)
Detect puppets by our creating them vs by nick suffixing. Not fool-proof at the moment as when/if an ircd forces a nick change before the listener sees it as the old nick it'll be a hanging nick and could double-ups or loops.

Assuming someone is already using nick suffixes this should work just like it did before but would likely fix #43 and fix #44.