Possibly fix relaying nick changes.
Noticed that it only captures the first nick change and then fails to
relay anything further for the nick and fails to capture further nick
changes. Think the call is held up when it sends into the channel and
never gets to the other "NICK" callbacks. Hopefully this will solve
that.