Better Support Private Messaging
This commit introduces the ability to private message in such a way as
to be considered "stable". Each bot get its own discriminator and this
makes it where if you were to connect 2 bots to the same Discord
identity but 2 different IRC networks when you send a private message
since the discriminator is required it will not accidentally be sent to
both IRC networks wherein both networks have a client with the same
nick. This also enhances privacy since it isn't guaranteed the same
person has the same nick on both networks.

Pending: log a fatal and exit if Discriminator == "".