Provide more helpful errors for invalid keys.
When it is first load, if an invalid key is present in `discord_format`
we inform the user of a fatal error and exit.
If it's on reload we don't apply the new format and inform the user such
was not done and that there is an error with an invalid key being
present in the `discord_format` setting.