Add `irc_puppet_prejoin_commands` (#88)
This adds support for interpolating the current nick of a connection so as to remove the need for hardcoded modes being set in the code. Provides for a separte list for the main connection (listener or the simple mode relay) and for puppets.

We discussed something akin to cascading previously of the prejoin onto the puppets where puppets would execute both sets of commands but the listener would only execute prejoin. I think it's best to keep these separate completely since I know inspircd allows for configurations where a user can set a mode on themself but cannot remove it. A possibly good example is mode +x being locked after set some places. So if the listener sets +x due to prejoin, and the connections set +x as well due to, but you want to -x later in the puppet_prejoin... you're stuck.

This hasn't been tested yet, and should supplant #70. This does not set any defaults whatsoever (perhaps a sane default of +D for puppet_prejoin makes sense?).

Co-authored-by: Qais Patankar <qaisjp@gmail.com>