Feature: Ignore messages containing matching text (#92)
This feature allows for when a message matches from Discord towards IRC
or the other way around for the matching message to be dropped
(filtered).

lets say there is an urban dictionary bot on xyz side of the bridge. Now it also does other useful things, but we know that urban dictionary can be floody for say "the longest definition on urbandictionary" query. Luck has it that whoever made the bot decided each line of urban dictionary related output should be prefixed "[ud]" so we can filter on this and keep the remainder of useful things. It also conveniently works for ignoring users that are behind a relay when a channel runs multiple kinds of relays